requirements.txt
slack_bolt
requests
google-cloud-firestore
flask
gunicorn
slack-sdk